refactoring folder step 2
