+++ src/composition-root.ts
@@ -10,9 +10,9 @@ import { BlogsQueryRepository } from './query-reositories/blogs-query-repository
 import { CommentsQueryRepository } from './query-reositories/comments-query-repository'
 import { PostsQueryRepository } from './query-reositories/posts-query-repository'
 import { UsersQueryRepository } from './query-reositories/users-query-repository'
-import { EmailAdapter } from './adapters/email-adapter'
-import { JwtService } from './application/jwt-service'
-import { EmailManager } from './managers/email-manager'
+import { EmailAdapter } from './infrastructure/email-adapter'
+import { JwtService } from './infrastructure/jwt-service'
+import { EmailManager } from './infrastructure/email-manager'
 import { AuthController } from './controllers/auth-controller'
 import { BlogsController } from './controllers/blogs-controller'
 import { CommentsController } from './controllers/comments-controller'
