+++ src/controllers/auth-controller.ts
@@ -8,7 +8,7 @@ import {
 } from '../models/auth-models'
 import { Request, Response } from 'express'
 import { HTTP_STATUSES } from '../constats/status'
-import { JwtService } from '../application/jwt-service'
+import { JwtService } from '../infrastructure/jwt-service'
 import { v4 as uuidv4 } from 'uuid'
 import { settings } from '../settings/settings'
 import { checkAuthorizationAndGetPayload } from './helper'
