+++ src/infrastructure/email-adapter.ts
