+++ src/middleware/authorization/bearer-auth.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Response, Request } from 'express'
-import { HTTP_STATUSES } from '../constats/status'
-import { jwtService, usersService } from '../composition-root'
+import { HTTP_STATUSES } from '../../constats/status'
+import { jwtService, usersService } from '../../composition-root'
 
 export const bearerAuth = async (req: Request, res: Response, next: NextFunction) => {
     if (!req.headers.authorization) {
