+++ src/routers/comments-router.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express'
-import { bearerAuth } from '../authorization/bearer-auth'
+import { bearerAuth } from '../middleware/authorization/bearer-auth'
 import { getUserIdByAccessToken, idValidation, inputValidation } from '../middleware/input-validation'
 import { contentInCommentValidation } from '../middleware/input-comments-validation'
 import { commentsController } from '../composition-root'
