+++ src/routers/users-router.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express'
-import { auth } from '../authorization/basic-auth'
+import { auth } from '../middleware/authorization/basic-auth'
 import { idValidation, inputValidation, queryValidation } from '../middleware/input-validation'
 import { emailValidation, loginValidation, passwordValidation } from '../middleware/input-users-validation'
 import { usersController } from '../composition-root'
